feat(webRTC): add button call
When we visit someone else page we can call him/her by press the "Call" button.
The current changes is still not good as we are not having a things like let us to search the users or friends on this context.
Since the previous changes, we have to type the user view url by hand like this one:
http://localhost:3000/home#/user/view/525404bb3d296ac819000003
This is really cumbersome.